requirements.txt
PyYAML==6.0.2
playwright==1.50.0
openai==1.61.0